Add release signing configuration
This checks if the keystore file is present to add the release signing configuration in order to still allow the debug build if the key isn't present.
The signing configuration values are first read from the signing.properties file, which is a local file that will not be pushed.
If this configuration file isn't here, we try to read the values from the gradle command line, which will be useful for the signing process by the CI.