Refactor the whole database.
We need to be able to execute several actions after a detection, not only a click or swipe.
To achieve this, we need to refactor the database.
We will now have Events instead of Clicks, which will contain a list of Conditions and a
list of Actions.

NB: Commits on this branch are not guaranteed to compile.